`gcc -Wall' points out some signed/unsigned comparisons. Fixed.
git-svn-id: svn://svn.tartarus.org/sgt/charset@4090 cda61777-01e9-0310-a592-d414129be87e

+++ sbcs.c
@@ -43,9 +43,9 @@ int write_sbcs(charset_spec const *charset, long int input_chr,
     while (i+1 < j) {
 	k = (i+j)/2;
 	c = sd->ucs2sbcs[k];
-	if (input_chr < sd->sbcs2ucs[c])
+	if (input_chr < (long int)sd->sbcs2ucs[c])
 	    j = k;
-	else if (input_chr > sd->sbcs2ucs[c])
+	else if (input_chr > (long int)sd->sbcs2ucs[c])
 	    i = k;
 	else {
 	    emit(emitctx, c);
